Create audio manager.
This audio manager is a singleton, making it easy to call its static instance and play an audio from any object. The audio files are stored in and controlled by the AudioManager class, making it easy to change the audio since it's in one place. It is also a facade for the audio player which plays sounds, and the mixer controller which updates volume controls.