Made scripts and SO's to handle object interaction
3 scripts were made to handle object interaction to divide responsibility between identifying interactable objects, modifying the objects, and tracking the desired states.
The InteractableObject script is attached to objects players can interact with to modify the object's outline.
The InteractableObjectSearcher script seeks out objects with the InteractableObject script with a raycaster to activate the specific highlight actions.
The InteractableSettingsSO script stores the default desired settings for object highlights/outlines. It was made to allow for different object highlights for various types of interactable objects.