Create a screen interactable
When interacted with, the view of the player is moved into the screen so that it can be seen better. This is done with a camera swapper class that smoothly transitions between two cameras. It also swaps the action map from player to UI, to set up interaction with the screen.