Add textures, materials, and models for main rooms
This adds texture maps including normal, base-color, metal and height maps for lobby, office, and terminals 1-5.
The objects added are static in prefabs so it allows for baked lighting. The objects do not have collision boxes by themselves, as this is dealt with through prefabs. Other objects including shelves, boxes, doors, and a test building machine.