Add scene object prefabs
This allows for drag and drop full rooms inside the project scenes. There are different prefabs for the different levels of corrosion, allowing for different prefabs to be used with different player statuses. There are two post process volume prefabs, one for when 'wired' after drinking coffee and one corroded for player status.