add render settings
This allows for the two post process volumes to actually operate using their respective post process profiles.
One for when drinking coffee and one for when coffee is ignored. These will go into settings because they are profiles and not actual game objects, just the settings that the post process volume prefabs will take on.