Add Dither Shader
This shader is used through a UI canvas that is connected to the main camera. This is not a world space shader. To have the shader work, It's output rendering raw image called Renderer, needs to be connected to a UI canvas. It also needs to be the output texture under the main camera, under the output tab in the cameras inspector. The actual shader material has multiple different dither textures, and editable Dither opacity as well as Quantization, which is the color steps.