Add item pick up and rotation
Player can now pick up and rotate objects with the build items script. If something is between the player and the object it will attempt to move closer to the player.
If the distance between the player and the interrupting object is too small, it will drop the object. The clear current interaction method was made public to facilitate this.
Two actions for X and Y rotations were added to the input system for consistency with other inputs.