Add objects to scene and lightmapping
This populates the lighting and post process test scene with a base level of all main rooms. There is also the addition of post process volumes in scene and working camera with shader attached. This also adds baked lightmaps to the objects in the scene and lighting data. This also updated the Clean Post Process Profile to be mapped to the volume in level instead of using the base standard profile.