Create Pickup Object Prefab and Test Scene
Object Manipulation test scene has an object to pickup and a wall to allow for thorough testing. Pickup Item prefab allows the seamless addition of build items to other scenes for testing or usage.
To avoid unintentional dropping of the item due to objects being on the ground, a tag was added to identify and ignore the ground for the dropping item code.