Change pickup logic to fix phasing and dropping
Most changes were made to fix the object automatically dropping when picked up and the object phasing through objects when it shouldn't be.
The item pickup now uses the original orientation and distance the object is grabbed at to handle the default dropping distance.
It takes the distance to move the object before dropping via a percentage of the distance to handle issues with raycasts to each corner being different lengths.
Additionally, the check for what scene objects can be ignored when the raycast hits them was generalized so it could be utilized more easily in the future.