Change Distance To Interactable Check
Since there was a discrepancy between the distance to the object found by the raycast hit distance and the distance magnitude, a single method was made to consolidate the logic for consistency.
This ensures the distance is the same when checking if an object can be interacted with and if it is too far away and should be dropped.
The scene was slightly modified during this check to streamline testing this by placing the object just at the edge of the interactable distance to see if it was still allowing interaction outside the proper distance.