Implement stop interact subscriber
The interable object searcher script now subscribes to the stop interact
void event channel so that any function can raise an event to clear the
current interaction. The build item script was then changed to raise
that event rather than needing an instance of the interactable object
seracher to decouple them.