Create a module interaction system
This includes a screen module template prefab that allows you to quickly set up a new module. It includes functionality for clicking on the screen.

There is also a button, and it should not be difficult to add more analogue controls.

The camera controls script was also updated to make mouse inputs possible while interacting with the module.