Slider add and lighting apply
Added slider object variations (right, middle, left) and created prefabs and textures for all including a mount for them to mount to a terminal. Prefabs made for each of their clean and corroded versions to drag and drop into the level. Re insured Lighting would always be applied to the scene and wouldn't re-bake lightmaps on startup.