Check for collisions with pickup valid placement
Previously we were using raycasts to check the corners of objects but
this had some downsides, mainly forcing us to use box colliders for
everything regardless of actual shape. Updated it to briefly enable the
collider to check for collisions and move it as needed.

Also moved the IgnoreItemCollision tag to a layer so that it can be used
as a layer mask for raycasting purposes.