Prevent pick up new item while holding an item
IInteractable now has its own function for confirming if it can be
interacted with or not which the object searcher uses. We're not raising
the GameObject the player picks up and if one has been set it won't
allowed interact so now you have to drop your current item to pick up a
new item.