Fix crouching in main office
Added a separate layer to ignore the post processes since they have box colliders to allow them to be easily moved and applied over the prefab rooms.
Modified the CheckCapsule part of the player controller script because the check was constantly interacting with the post process box colliders.
Set the ignored items and post processes to a list of layers to ignore. Also added the player layer as the player was colliding with themselves somehow and couldn't uncrouch.