Add mover script to move objects
Created a script to automatically move an object from one place to another over a specified period of time. Also can move back to the original position. This was primarily for doors, but was generalized as a mover script to work with anything.
Created a DoorMover specific script that works for automatic doors via a door trigger that moves it to a new position when entering the trigger zone and returns to its original position when leaving the trigger zone.
Created Scriptable Objects for these actions to work with the event subscriber system within the main scene.