Make prefabs work with doors
Created a base prefab template for the doors and modified the main office room and elevator prefabs to incorporate the new doors prefab. Removed the static part of the doors to allow them to move.
Additionally modified the player prefab, adding a rigidbody and capsule collider to work with the door triggers.
Small modifications to the player prefab were also made to allow the dithering effect and post process to be visible while the player moves through the scene.