Security and Coffee add
Added coffee cup and carafe. Both have respective materials and are inside the prefab CoffeeComponents. *Note* Not baked into lighting, using realtime lights, can be made static later though. Added Speaker for transmissions to reach the player, with its own prefab and added to scene. Also not baked and using realtime lighting with a spot light inside the prefab. Added security cameras to scene with their own prefab, along with a script called SecurityCameraFollow that allows them to follow a player when dragging the player camera into the open space in the inspector. under the script which is attached to the SecurityCameraBase object in the prefab.