Began changes to work with dither camera
Started making modifications to the scene, prefabs, and various scripts that used to work with the main camera to swap it to a DitherCamera prefab that can be applied to all the various minigame cameras.
Made some small edits to the scene to speed up door movement and fix the door triggers. Also added a player tag to the player.
Further scripts like the camera swapper and others still remain unchanged but will need to be changed moving forward to work with the new camera set up.