Made placement location more intuitive to set
Since the placement location was a child of other objects setting up the
placement location was unintuitive because it needed to be set to the
scale of the object being placed there. This new script ignores the
actal transform and draws a box where your expected item will be go
based on its scale and then on awake change the transform to match
that.