CRT Screen Shader Add
This shader utilizes Unity's shader graph, and will allow for each monitor screen to be customized with this CRT shader. There is a lit material derived from this shader so it will be applied to each plane that is a "screen" in the scene. There are a few fields that can be filled in the materials inspector. The first being the Render texture area which should be whatever the given screen should be rendering and using that specific minigame render texture. The other fields are self explanatory and contain the brightness, line texture, warp texture and tiling, and emissive color.