UI Component Add
In this there are post-it looking buttons to be used as UI elements on a pause menu, and two whiteboards to be used as the backdrop of these post-it buttons during pause and settings menus. These physical elements all have prefabs, and the buttons have a UI canvas attached to them and a TMP button to be assigned as well. The button has editable text to be matched to it's assignment under it. *Note* These objects also have a box collider attached as well as a script called UIMovement that moves the object up when mouse hovers and scales and rotates it when clicked, which operates separate from the UI-button and what it will be assigned to with it's on click operation. There are four post-it button prefabs. All objects have textures assigned to them. These buttons also use a new font that has been added. There are other UI element prefabs that have been added including scroll buttons, a slider, and an input field for inputting controls, sensitivity, brightness, etc.