Height Adjustments to Doors/Hallway
Raised the ceiling in the hallway to match the ceiling height of the
main room, because they were different originally. This helps with the
updated player height, as the player can now get through all doors fine.
The walls of these rooms had to be scaled higher (from 1 to 1.4) because
the door arch in the wall was still contacting the player's collider.