Organize files and generalize screen module interaction start
The prefabs and SO folders weren't updated to fit the new organization structure in the last commit, so that's fixed now. Also, there's now only one start module interact SO. This technically means that if a control for another screen module was in the corner of the players screen, they could click it.