Cursor and UI base add
This adds the base components of ui, including boards, sliders, and buttons, that can be renamed and tied to different functions. These are in separate prefabs and can be organized into set boards for the pause menu with the pause menu blockout in slack as a blueprint. This also adds multiple cursors to use when hovering and clicking different objects, and in different stages of the game, along with a crosshair.