Add PlayerTransform Event Channel/Subscriber
To decrease the coupling of playerController, we use the
PlayerTransformEventChannel to move the player to the correct spot after
they finish interacting with a terminal. I added the
PlayerTransfromPayload.cs as a struct to store all the values the
playerController was previously handling. It is a more dynamic system
now.