Implement EventChannel in PlayerController
The PlayerController has been updated to include a OnEnable and
OnDisable to handle the events of the PlayerTransformEventChannel. It
will call a new function HandleMovePlayer, which calls
SetFootPositionAndRotation. This rotation was added to be sure the
camera ends up in the right spot after interacting. The ScreenModule.cs
changes were made to fit with the PlayerTransform Struct to update the
player's position, rotation, and camera rotation.