Camera animation change and rename
Instead of using a script for the camera movement it is now using a looping animation to give the breathing effect. this is inside assets/art/animations. There is also renaming including correcting spelling. There are some numbers kept in assets because in later pull requests there will be different variations. These assets include Shell1 and ArmFoot1 along with the material CoffeeSpill2.