Add Main Menu Scene
This adds a main menu scene to be used after the splash screen. There is a prefab of the whole scene under the folder SceneObjects. This contains a screen space canvas with buttons that have not been tied to other scripts and scenes yet. There is an animation and animation controller tied to the main camera to create a breathing and slight moving effect. This scene has it's own baked light maps inside of assets/art/lighting/mainmenu. The dither shader has already been applied in the scene and the only thing to add is to connect the buttons to functions.