Add Loading Screen
This loading screen is it's own scene. There are no scripts attached yet. There is a slider that can be tied to the loading status via a script. To operate this slider the value can be found in the inspector of the slider under the canvas. There is a shader tied to the animated background. There is no dither shader applied but can be later on.