Update Audio Manager and Add Tag/AudioListener
The audio manager now has all the clips stored in its prefab with the
clip name matching the type name provided in SoundType.cs. For the
player to hear the sounds, I attached an AudioListener component to the
player prefab. An additional change to the MainRoom prefab and TagManager
was adding a "Ground" tag and applying that tag to the ground objects in
the MainRoom Prefab in order for objects to properly play the drop sound
when colliding with the ground.