Add Footsteps and Remove Echoing
Footsteps have been implemented through a coroutine in the
PlayerState.cs file. The walk and sprint states access this coroutine
whenever the player moves. There is a landing sound also being applied
in the JumpState.cs file when the player is grounded. Lastly, to fix an
echoing footstep problem, the objectPool needed one line of code added
(set velocity to 0) so it doesn't continuously echo around the player
after 1 step has been taken.