Implementing SFX for Doors and Items
The DoorMover script now plays an open/close sound when the doors are
opened/closed. Pickup Item has two cases of where the player can drop
the item, and whenever the item is dropped (from collision or hiting the
ground) it will play the drop item sound. Lastly, when an object is
placed correctly, in PlacementTrigger.cs, a sucessful locking sound will
play.