Create a box explosion effect
The effect is triggered when the player slams the box into the ground. It is fragmented using OpenFracture into a bunch of different game objects. The items in the box are set by the serialized BoxItemsSO.

The Box2 material was altered to change it to allow read/write since OpenFracture requires it.