Update Scripts and Landing Bug
To try to fix the landing bug, I moved the audio to play only when the
player leaves the jump state since IsGrounded was not working some of
the time. I still sometimes can get the bug to happen once, but I do not
think it is a priority at this time. The pickup script was modified for
easier logic. The player SFX scene now looks how the office would.