Make Modules Use DitherCam as Serialized Field
Originally, dithercams had a tag of mainCamera and that was how the
cameras were accessed in code files. These code files have been updated
to grab the dither cam of each terminal with a serialized field. The
prefabs have been updated to take the position of the dither cam as
well. Thanks for the help Tavin!