Fix box exploding mid-air
Before, it triggered anytime the item was dropped. Now, it triggers only when the item is dropped because it moved to an invalid position. This is essentially a collision.

The box exploder now uses an update function instead of a coroutine. This simplifies the code and removes the need for an event to trigger the coroutine. Since there can only be a few boxes in the scene, the performance hit will be negligible.

A logic error was fixed in OpenFracture.