Create build item constructor
Created a script that will automatically construct all the build items
with their placement nodes and save them as prefabs. Saves time of
manually placing all of them. Manual changes still need to be made to
set BuildItemRenamer or changing the placement trigger since that
defaults to the placement container location.