Add steam particle effect
steam particles have been added for when the player makes coffee, this effect is a particles system that is in the prefabs folder. This uses a shader graph and can be editable. This prefab can be dragged into the main level and should be in the correct position. It is not attached to the coffee being poured yet however. It's runtime can be changed in the inspector but is bound to 2.5 seconds as of now.