Steam Remake
In this the steam is remade and is more simplified and steam like. this discards a shader and instead just uses a material with one sprite. The steam prefab is in scene objects and the steam sprite is under the sprites folder. It has been attached to the coffee machine prefab and is in the main game scene. It's currently set to looping and can be toggled on and off as an object.