Add Scripting Logic and Cursor adjustments
Added a file for minigames scripts, a folder specifically for mazes. In
this folder, all the walls on the mazes will have the WallCollision
script attached, that is checking for an oncollision with the cursor.
The maze manager is mainly used to shuffle through mazes. The goal
trigger is to end the maze minigame.