Main Scene Lighting rework
Changed the lighting in the main scene from baked lighting to real time lighting. This ensures that held objects and other moving objects in the scene can be moved around and interact with lighting and not having to be baked into a lighting map. This also uses point lights and spot lights, since aera lights can only be baked, and have no real-time effect.