Make working UI scene with prefab menus included
The UI menu was set up with the basic functional menus and scripts attached as desired with everything saved in prefabs for easy modifications.
The UI scene should have all the basic screens that are used throughout the game within it and they are currently being controlled by simply turning the screen game objects on and off depending on when they are needed for simplicity.
Additionally, the MainMenu prefab and MainRoom prefab were modified to work with this UI scene and each screen was made into its own prefab so it can be worked with and modified simply.