Add animations and sprites for UI elements
Animations and specific sprites were created and chosen for the various buttons and UI elements ot make them visuallya pealing and interactive feeling to use.
All the sticky notes have really basic animations to make it easily distinguishable that the button has been interacted with.
Since we are using UI as it's simpler than game objects, we had to take images of the 3D sticky note models and turned those into sprites, so they are stored currently in a their own little subsection of the sprites art folder.