Modify settings and scripts for display changes
Various settings had to be changed to account for the display settings we change within the UI scripts via the buttons, sliders, and dropdowns.
Some settings needed modifications to the actual build or game settings so that things like the quality of the game could be modified or changed with the dropdown menu in the display panel.
Also, the AudioMixer required the master volume to be set so that it could be modified.
The input system had a number of unnecessary controls that added complexity and confusion that was more easily avoidable by removing all controls besides keyboard controls.