Fix variable consistency and functions in scripts
Made small modifications to remove unnecessary variables while also adding functionality to some scripts to make them work with the game loading and player preferences.
These scripts mostly had really small things that needed to be changed, such as changing from start to awake for timing, removing or modifying variables to be usable with changes that actually apply to the game and are saved.