Change mesh settings to read/writeable
This fixes an issue where the game cannot be additively loaded and played if the game object's mesh settings are not set to read write due to conflicts with the outline 3rd party script.
This comes from an issue where unity tries to combine meshes of static objects, but this causes problems with the oultine.
This is most easily fixed by making the mesh for them read/writable and avoiding static objects when possible.