Change scene appearance and functionality
I changed all the prefabs to work with the new button animations and keybindings. I also added a panel to confirm a player wanted to return to the main menu.
Additionally, the coffee cups mesh was changed and had to be fixed inside the main menu screen prefab as well.