Add/Remove Scripts to Terminal Prefabs
I added the ScreenTerminalLinker script to all terminal prefabs so other
people will not struggle on the aspect ratio for other minigames. In
making the CursorController abstract, it now gives warnings whenever it
is attached to game objects, so I made sure to remove the
cursorController script from the terminals to help avoid some of those
warnings, especially for our main scene. I did not remove the
cursorcontroller scripts from the test scenes.