Add Events for Maze Specific Events
The MazeCollsiion SO is for colliding with walls, which then uses
MazeManager to reset the position. MazeComplete is used when the goal is
achieved, then loads a new maze and turns the screen off through maze
manager. The mazeJoystick move was to be rid of the joystick test
channel. StartMazeInteract is used in place of StartModuleInteract.