Modify Scripts Based on Feedback.
The mazeCursor and cursor controller scripts have been modified so
cursorcontroller is now an abstract class that mazecursor inherits from.
Most of the functionality works the same, removed a big chunk of code
from CursorController as it only applied to mazeCursor. GoalTrigger and
WallCollisions have been modified to handle events. MazeManager.cs has
implemented TurnScreenOff as a method that takes in a bool, instead of
having two seperate functions. Scene has minor changes.: