Modify Code to Requested Changes
The CursorController script has been renamed to ScreenModuleCursor. The
screenmodule template prefab has removed the cursorcontroller script
from it's cursor. The mazeCollision script applies the logic from
wallCollision and goalTrigger in one script. The mazeManager added a
function for this to call StopMaze. All other changes are minor, fixing
other scripts that had CursorController as a type and changed to
ScreenModuleCursor. Comments for clairty have been added to
screenCameraLinker.