Add maze minigame art
this commit adds the background artwork for the maze minigame. This adds modular sprites to use for the minigame and uses a grid and tilemap to created the different levels of the minigame. This allows for editable levels and different situations. There are two shaders being used, one for the background and one for the foreground. The prefab is under the SceneObjects folder. There area two render textures being utilized, one to render the base image, then another to render the shader effects ontop, and that gets rendered onto the screen that the CRT shader is present on. the prefab can be referred to this. This also means that there are two cameras being used by these render textures.