Add minigame player pawn
this adds a drill machine for the player to control and a compartment of radiative material to go under/ out the back of the drill for deposit when the game is done. This also changes some file locations and it changes the base color texture that is used by the atmospheric converter buildable.