Add Siren
This adds two sirens, one clean and one corrode. These both have prefabs inside of SceneObjects, and can be dragged into the scene readily. These have a script attached to each of them that allows them to be toggled from active to inactive, and can be tied to different events. This script triggers an animation that causes the lights to turn on and spin.