Revamp the box items SO
It now has a count and a game object. This way, if there's more than one of each part, it will be easier to set that up. Also, there is now an explosiveness variable. Since it's powered by the objects spawned in the box colliding with eachother, it can be a little finicky and should be tweaked based on the items inside of it.