Add splash screen and load screen
Added the splash screen and loading screen to the UIMenus scene so you start on the splash screen and use the loading screens as a transition.
This also solved the need for delays while transitioning between different cameras when no camera rendering occurred.
Made a few bug fixes encountered during testing as well such as time being frozen and the player being unable to move.