Create a box to place the final build in
The box opening and closing animations are triggered by the build finished and build placed in box events.

Build order enforcer was changed so that it triggers an event when the build is finished. It also now properly checks the children for placement triggers.

An option to scale the entire object being placed was added to the placement trigger. This made it so the entire build could be placed in the box. There is also a new public function to set the required item name.