Add Planet State Viewer
This allows for the player to see the state of the planet and promotes urgency with a countdown of planet degradation. There are five sprites added starting with healthy and ending with destroyed. These all have been placed into an animator controller where the state is controlled by an integer. This integer can be tied to a time script or can be removed and changed to Booleans or something else. For now the basic state is the healthy planet and the animation is a non-reversible track to the destroyed planet. The prefab contains the animations applied to the sprite, and a camera rendering out a render texture to be applied to a screen with the CRT shader.