Create chutes for sending and recieving items
The in chute sends out items in a queue that is going to be set by the level loader. The last item in the queue will always be a box that can have something placed in it.

The out chute only opens when the build is placed in the out box, and the out chute will only accept the out box. Once the out box is placed in the chute, it raises the Win Level event.

This commit also includes testing materials, which can be used in the Chutes scene.