Ending Screen Add
This adds two ending screens, A fail and a success screen. These both use the same animation and are separate scenes to be loaded after either failing or succeeding. They have also been made into prefabs under EndingScreens folder in the Prefabs folder. This also changes the way that the animation of planet status is read. This uses a float instead of an integer, and switches on depending the level of the planet degredation in steps of 20, of if it is greater than, counting down from 100.