Fix animations
Removed the static state and bool which set it from all the option menu buttons and instead go to the disabled state by making it not interactable.
Replaced old code in OptionsMenu that set that to one that invokes the onClick function for the button and connected menu that was last used.
Deleted an extra line from keyrebinding.