Create script to swap corrode levels
A script can now be attached to all items that change materials to a
clean, corrode, or ultra corrode. It saves the given materials in a
dictionary and only swap to the highest level closest to whichever one
is set. So if you set to ultra corrode and it doesn't have one it will
set the corrode instead.