Clean up options menu and sprites folder
Removed the extra unused sticky notes and poster from the UISprites folder. Also moved the sticky notes and board to the UI folder under sprites to keep everything together.
Deleted the options menu as all the functionality was moved to be handled via the buttons' onClick functionalities. Added a prefab for those buttons
Small bug fix where the message for keyrebinding wasn't set to run when timescale was 0, so now it waits for realtime seconds. Also removed unnecessary transition in pause menu animator and fixed one more slider to add slider value changed in inspector.