Add Needed Scripts for MainTerminal
The individual screen managers, message, instruction, and planet, are
all apart of the IScreen interface for activating and deactivating the
screens. The only way to view the canvas with the CRT shader was with
the each Canvas viewer, who uses the render texture of the output of the
viewcamera, then the screen gets that and applies a CRT over it.