Make scripts for minigames and minigame manager
Made a minigame abstract class that can be used by all minigames with basic functions for StartMinigame, StopMinigame, ResetMinigame, and a variable to get the duration of time the minigame took.
Also made a minigame manager that takes a list of minigames and turns them on at times based on the minigameStatsSO. The SO holds the information for which terminal siren to turn on and how long there should be between alarms as well as how much damage to affect the planet with.
Made a simple planet health tracking script mostly as a test for the moment so a slider can be added and checked to see if the planet health is properly deteriorating. This should later be modified to change the health of the planet on the main terminal screen.
Additionally converted maze manager to be a subclass of minigame, reworking the logic to work with that, and also modified ScreenModule so it only allows the terminal to be interactable if the minigame on it is going off.